requirements.txt
opencv-python
jupyter
ipykernel
matplotlib
numpy
pillow
tensorflow <= 2.5
tensorflow_datasets
